requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.8
iniconfig==2.0.0
jiter==0.5.0
numpy==2.1.0
openai==1.42.0
packaging==24.1
pluggy==1.5.0
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
slack_bolt==1.20.1
slack_sdk==3.31.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2